+++ pcap-linux.c
@@ -339,12 +339,6 @@ static int pcap_setfilter_linux(pcap_t *, struct bpf_program *);
 static int pcap_setdirection_linux(pcap_t *, pcap_direction_t);
 static int pcap_set_datalink_linux(pcap_t *, int);
 static void pcap_cleanup_linux(pcap_t *);
-#ifdef PCAP_SUPPORT_ODP
-static void pcap_odp_init(pcap_t *);
-static int pcap_activate_odp(pcap_t *);
-static int pcap_read_odp(pcap_t *, int, pcap_handler, u_char *);
-static void pcap_cleanup_odp(pcap_t *);
-#endif
 
 union thdr {
 	struct tpacket_hdr		*h1;
@@ -6116,292 +6110,3 @@ reset_kernel_filter(pcap_t *handle)
 				   &dummy, sizeof(dummy));
 }
 #endif
-
-#ifdef PCAP_SUPPORT_ODP
-pcap_t *
-odp_create(const char *device, char *ebuf, int *is_ours)
-{
-	pcap_t *handle;
-	struct pcap_odp *podp;
-
-	*is_ours = (!strncmp(device, "odp:", 4)
-		    || !strncmp(device, "netmap:", 7)
-		    || !strncmp(device, "vale", 4));
-	if (! *is_ours)
-		return NULL;
-	if (!strncmp(device, "odp:", 4)) {
-		handle = pcap_create_common((device + 4), ebuf, sizeof(struct pcap_linux));
-		handle->selectable_fd = -1;
-		podp = handle->priv;
-		podp->is_netmap = false;
-#ifdef ODP_HAVE_NETMAP
-	} else if (!strncmp(device, "netmap:", 7)) {
-		handle = pcap_create_common((device + 7), ebuf, sizeof(struct pcap_linux));
-		handle->selectable_fd = -1;
-		podp = handle->priv;
-		podp->is_netmap = true;
-#endif
-	} else {
-		handle = pcap_create_common(device, ebuf, sizeof(struct pcap_linux));
-		handle->selectable_fd = -1;
-		podp = handle->priv;
-		podp->is_netmap = false;
-	}
-	if (handle == NULL)
-		return NULL;
-
-	handle->activate_op = pcap_activate_odp;
-	return (handle);
-}
-
-static void
-pcap_odp_init(pcap_t *handle)
-{
-	odp_buffer_pool_t pool;
-	odp_pktio_t pktio;
-	void *pool_base;
-	char inq_name[ODP_QUEUE_NAME_LEN];
-	odp_queue_t inq_def;
-	odp_queue_param_t qparam;
-	int fd;
-	int ret;
-	struct pcap_odp *podp;
-
-	/* Init ODP before calling anything else */
-	if (odp_init_global()) {
-		fprintf(stderr, "Error: ODP global init failed.\n");
-		exit(EXIT_FAILURE);
-	}
-
-	/* Create thread structure for ODP */
-	if (odp_init_local()) {
-		ODP_ERR("Error: ODP local init failed.\n");
-		exit(EXIT_FAILURE);
-	}
-
-	/* Is pool have been created in another theard ? */
-	pool = odp_buffer_pool_lookup("packet_pool");
-	if (pool == ODP_BUFFER_POOL_INVALID) {
-		/* Create packet pool */
-		pool_base = odp_shm_reserve("shm_packet_pool",
-				SHM_PKT_POOL_SIZE, ODP_CACHE_LINE_SIZE, 0);
-		if (pool_base == NULL) {
-			fprintf(stderr,
-				"Error: packet pool mem alloc failed.\n");
-			exit(EXIT_FAILURE);
-		}
-
-		pool = odp_buffer_pool_create("packet_pool", pool_base,
-				SHM_PKT_POOL_SIZE,
-				SHM_PKT_POOL_BUF_SIZE,
-				ODP_CACHE_LINE_SIZE,
-				ODP_BUFFER_TYPE_PACKET);
-		if (pool == ODP_BUFFER_POOL_INVALID) {
-			fprintf(stderr, "Error: packet pool create failed.\n");
-			exit(EXIT_FAILURE);
-		}
-		odp_buffer_pool_print(pool);
-	} else {
-		fprintf(stdout, "packet pool have been created.\n");
-	}
-
-	/* Open a packet IO instance for this thread */
-	/* if any device, need ODP support */
-
-	podp = handle->priv;
-
-	podp->pktio = odp_pktio_open(handle->opt.source, pool);
-	if (podp->pktio == ODP_QUEUE_INVALID) {
-		fprintf(stderr, "  Error: pktio create failed %s\n", handle->opt.source);
-		return;
-	}
-
-	/*
-	 * Create and set the default INPUT queue associated with the 'pktio'
-	 * resource
-	 */
-	qparam.sched.prio  = ODP_SCHED_PRIO_DEFAULT;
-	qparam.sched.sync  = ODP_SCHED_SYNC_ATOMIC;
-	qparam.sched.group = ODP_SCHED_GROUP_DEFAULT;
-	snprintf(inq_name, sizeof(inq_name), "%i-pktio_inq_def",
-		 (int)podp->pktio);
-	inq_name[ODP_QUEUE_NAME_LEN - 1] = '\0';
-
-	inq_def = odp_queue_create(inq_name, ODP_QUEUE_TYPE_PKTIN, &qparam);
-	if (inq_def == ODP_QUEUE_INVALID) {
-		fprintf(stderr, "  Error: pktio queue creation failed\n");
-		return;
-	}
-
-	ret = odp_pktio_inq_setdef(podp->pktio, inq_def);
-	if (ret != 0) {
-		fprintf(stderr, "  Error: default input-Q setup\n");
-		return;
-	}
-
-	printf("  created pktio:%02i, queue mode\n"
-		"  default pktio%02i-INPUT queue:%u\n",
-		podp->pktio, podp->pktio, inq_def);
-
-}
-
-static int
-pcap_activate_odp(pcap_t *handle)
-{
-	struct pcap_linux *handlep = handle->priv;
-	const char	*device;
-	int		status = 0;
-	int		arptype;
-	struct ifreq	ifr;
-
-	/* initial ODP stuff */
-	pcap_odp_init(handle);
-
-	device = handle->opt.source;
-
-	handle->inject_op = pcap_inject_linux;
-	handle->setdirection_op = pcap_setdirection_linux;
-	handle->set_datalink_op = pcap_set_datalink_linux;
-	handle->setnonblock_op = pcap_setnonblock_fd;
-	handle->getnonblock_op = pcap_getnonblock_fd;
-	handle->cleanup_op = pcap_cleanup_odp;
-	handle->read_op = pcap_read_odp;
-	handle->setfilter_op = pcap_setfilter_linux;
-	handle->stats_op = pcap_stats_linux;
-
-	/*
-	 * The "any" device is a special device which causes us not
-	 * to bind to a particular device and thus to look at all
-	 * devices.
-	 */
-	if (strcmp(device, "any") == 0) {
-		if (handle->opt.promisc) {
-			handle->opt.promisc = 0;
-			/* Just a warning. */
-			snprintf(handle->errbuf, PCAP_ERRBUF_SIZE,
-				 "Promiscuous mode not supported on the \"any\" device");
-			status = PCAP_WARNING_PROMISC_NOTSUP;
-		}
-	}
-
-	handlep->device = strdup(device);
-	if (handlep->device == NULL) {
-		snprintf(handle->errbuf, PCAP_ERRBUF_SIZE, "strdup: %s",
-			 pcap_strerror(errno));
-		return PCAP_ERROR;
-	}
-
-	/* copy timeout value */
-	handlep->timeout = handle->opt.timeout;
-
-	/*
-	 * If we're in promiscuous mode, then we probably want
-	 * to see when the interface drops packets too, so get an
-	 * initial count from /proc/net/dev
-	 */
-	if (handle->opt.promisc)
-		handlep->proc_dropped = linux_if_drops(handlep->device);
-
-	/* + activate_new */
-	/* Will create a sock_fd just for setting */
-	status = activate_new(handle);
-	if (status < 0)
-		goto fail;
-
-	/* Allocate the buffer */
-	status = 0;
-	if (handle->opt.buffer_size != 0) {
-		/*
-		 * Set the socket buffer size to the specified value.
-		 */
-		if (setsockopt(handle->fd, SOL_SOCKET, SO_RCVBUF,
-			       &handle->opt.buffer_size,
-		    sizeof(handle->opt.buffer_size)) == -1) {
-			snprintf(handle->errbuf, PCAP_ERRBUF_SIZE,
-				 "SO_RCVBUF: %s", pcap_strerror(errno));
-			status = PCAP_ERROR;
-			goto fail;
-		}
-	}
-
-	handle->buffer = malloc(handle->bufsize + handle->offset);
-	if (!handle->buffer) {
-		snprintf(handle->errbuf, PCAP_ERRBUF_SIZE,
-			 "malloc: %s", pcap_strerror(errno));
-		status = PCAP_ERROR;
-		goto fail;
-	}
-
-	handle->selectable_fd = handle->fd;
-	/* - activate_new */
-
-	return status;
-
-fail:
-	pcap_cleanup_linux(handle);
-	return status;
-}
-
-static int
-pcap_read_odp(pcap_t *handle, int max_packets, pcap_handler callback,
-	      u_char *userdata)
-{
-	odp_packet_t pkt;
-	odp_buffer_t buf;
-	u_char *bp;
-	struct pcap_linux *handlep = handle->priv;
-	struct pcap_pkthdr pcap_header;
-	struct timeval ts;
-	long n = 1;
-	struct pcap_odp *podp = handle->priv;
-
-	for (n = 1; (n <= max_packets) || (max_packets < 0); n++) {
-		/* Use schedule to get buf from any input queue */
-		buf = odp_schedule(NULL, ODP_SCHED_WAIT);
-
-		pkt = odp_packet_from_buffer(buf);
-		if (odp_unlikely(odp_packet_error(pkt)))
-			goto clean_buf; /* Drop */
-
-		/* fill out pcap_header */
-		gettimeofday(&ts, NULL);
-		pcap_header.ts = ts;
-		bp = odp_packet_l2(pkt);
-		pcap_header.len	= odp_packet_get_len(pkt);
-		pcap_header.caplen = pcap_header.len;
-
-		/* ODP not yet support filtering_in_kernel */
-		if (handlep->filter_in_userland && handle->fcode.bf_insns) {
-			if (bpf_filter(handle->fcode.bf_insns, bp,
-				       pcap_header.len,
-				       pcap_header.caplen) == 0) {
-				/* rejected by filter */
-				n--;
-				goto clean_buf;
-			}
-		}
-
-		callback(userdata, &pcap_header, bp);
-
-clean_buf:
-		handlep->packets_read++;
-		odp_buffer_free(buf);
-
-		if (handle->break_loop) {
-			handle->break_loop = 0;
-			return PCAP_ERROR_BREAK;
-		}
-	}
-
-	return max_packets;
-}
-
-static void
-pcap_cleanup_odp(pcap_t *handle)
-{
-	struct pcap_odp *podp = handle->priv;
-
-	odp_pktio_close(podp->pktio);
-	pcap_cleanup_linux(handle);
-}
-#endif /* PCAP_SUPPORT_ODP */
